Second commit
Implemented BS4 to retrieve the correct link for each job.

Fixed issue where the dynamically generated job classes were becoming stale if the user searched an additional time.

Created a function to search for salary information in the description of the job posting(currently WIP).

Placed a file check so that each time the script is ran it does not overwrite the file if it already exists.  It simply appends, if the file does not exist, it creates.

Updated requirements.txt for the experimental branch.

POTENTIAL ISSUES:

Introduced an issue where appending to a file also carries over the columns names AND jobs( I only want the job information).

Using BS4 & Selenium to parse the job posting to retrieve more information(desc, salary, actual job link, etc) has increased the runtime of the script to complete its operation.